refactor: extract keyword_rerank, remove dead code
Shared keyword_rerank() in hybridSearch.ts replaces 3 inline copies of
the keyword-search + RRF-merge + pool-reorder pattern. Callers in
chat.ts (decomposition and fallback paths) and notes.ts (panel) each
reduced to one function call.

Also removes: dead original_body save/restore in chat decomposition
(no longer mutates note.body after MaxSim refactor), verbose fallback
debug logging, unused imports.
