+++ src/commands/chat.ts
@@ -4,7 +4,7 @@ import { BlockEmbedding, NoteEmbedding, extract_blocks_links, extract_blocks_tex
 import { update_panel } from '../ux/panel';
 import { get_settings, JarvisSettings, ref_notes_prefix, search_notes_cmd, user_notes_cmd, context_cmd, notcontext_cmd } from '../ux/settings';
 import { split_by_tokens, preprocess_query, clearApiResponse, clearObjectReferences, stripJarvisBlocks } from '../utils';
-import { decompose_query, keyword_search_chunks, rrf_merge, maxsim_search } from '../notes/hybridSearch';
+import { decompose_query, keyword_rerank, maxsim_search } from '../notes/hybridSearch';
 import { getLogger } from '../utils/logger';
 
 const log = getLogger();
@@ -233,7 +233,6 @@ async function get_chat_prompt_and_notes(
       }
     }
     // LLM query decomposition (when enabled, skip if user set explicit context)
-    const original_body = note.body;
     let nearest: NoteEmbedding[] = [];
     let decomposed = false;
     if (settings.notes_decompose_query && !prompt.search && prompt.notes.size === 0 && !prompt.context) {
@@ -257,28 +256,9 @@ async function get_chat_prompt_and_notes(
           const scored = maxsim_search(query_embeddings, sub_embeds, cache, note.id, settings);
 
           if (scored.length > 0) {
-            nearest = [{id: note.id, title: 'Chat context', embeddings: scored, similarity: null}];
-
-            // keyword merge (RRF - different scoring system)
             const keywords = sub_queries.flatMap(sq => sq.keywords).filter(k => k.length > 0);
-            if (settings.notes_keyword_weight > 0 && keywords.length > 0) {
-              const seen = new Set<string>();
-              const kw_chunks: BlockEmbedding[] = [];
-              for (const kw of keywords) {
-                for (const chunk of await keyword_search_chunks(kw, scored, 100)) {
-                  const key = `${chunk.id}:${chunk.line}`;
-                  if (!seen.has(key)) { seen.add(key); kw_chunks.push(chunk); }
-                }
-              }
-              if (kw_chunks.length > 0) {
-                const semantic_top = scored.slice(0, settings.notes_max_hits);
-                const merged = rrf_merge(semantic_top, kw_chunks,
-                  settings.notes_max_hits, settings.notes_keyword_k, settings.notes_keyword_weight);
-                const merged_keys = new Set(merged.map(b => `${b.id}:${b.line}`));
-                const tail = scored.filter(b => !merged_keys.has(`${b.id}:${b.line}`));
-                nearest[0].embeddings = [...merged, ...tail];
-              }
-            }
+            const reranked = await keyword_rerank(scored, keywords, settings);
+            nearest = [{id: note.id, title: 'Chat context', embeddings: reranked, similarity: null}];
             decomposed = true;
           }
         }
@@ -286,8 +266,6 @@ async function get_chat_prompt_and_notes(
     }
 
     if (!decomposed) {
-      // restore original context (decomposition may have mutated note.body)
-      note.body = original_body;
       // existing path: single find_nearest_notes + optional keyword merge
       nearest = await find_nearest_notes(sub_embeds, note.id, note.markup_language, note.title, note.body, model_embed, settings, false);
       if (nearest.length === 0) {
@@ -295,37 +273,11 @@ async function get_chat_prompt_and_notes(
       }
 
       // keyword merge (when enabled, without decomposition)
-      if (settings.notes_keyword_weight > 0 && !prompt.search && prompt.notes.size === 0) {
+      if (!prompt.search && prompt.notes.size === 0) {
         const keyword_source = prompt.last_user_prompt || note.title || '';
         const keyword_query = preprocess_query(keyword_source).slice(0, 200).trim();
         if (keyword_query.length > 0) {
-          const keyword_chunks = await keyword_search_chunks(
-            keyword_query, nearest[0].embeddings, 100);
-          if (keyword_chunks.length > 0) {
-            const semantic_top = nearest[0].embeddings.slice(0, settings.notes_max_hits);
-            const merged = rrf_merge(semantic_top, keyword_chunks, settings.notes_max_hits, settings.notes_keyword_k, settings.notes_keyword_weight);
-
-            if (settings.notes_debug_mode) {
-              const sem_keys = new Set(semantic_top.map(b => `${b.id}:${b.line}`));
-              const merged_key_set = new Set(merged.map(b => `${b.id}:${b.line}`));
-              const promoted: string[] = [];
-              const demoted: string[] = [];
-              for (const key of merged_key_set) {
-                if (!sem_keys.has(key)) { promoted.push(key); }
-              }
-              for (const key of sem_keys) {
-                if (!merged_key_set.has(key)) { demoted.push(key); }
-              }
-              log.info(`[Hybrid] query: "${keyword_query.slice(0, 80)}"`);
-              log.info(`[Hybrid] semantic: ${semantic_top.length}, keyword: ${keyword_chunks.length}, merged: ${merged.length}`);
-              if (promoted.length > 0) { log.info(`[Hybrid] promoted by keyword: ${promoted.join(', ')}`); }
-              if (demoted.length > 0) { log.info(`[Hybrid] displaced from semantic: ${demoted.join(', ')}`); }
-            }
-
-            const merged_keys = new Set(merged.map(b => `${b.id}:${b.line}`));
-            const tail = nearest[0].embeddings.filter(b => !merged_keys.has(`${b.id}:${b.line}`));
-            nearest[0].embeddings = [...merged, ...tail];
-          }
+          nearest[0].embeddings = await keyword_rerank(nearest[0].embeddings, [keyword_query], settings);
         }
       }
     }
